FIXes OSX > 10.4 num_Cpus() the good and posix way
sysconf() is  part of IEEE Std 1003.1-1988 (POSIX.1).
Fortunately, it *is* available to OSX `>=10.5`.
At least on `~10.14`, the old unposix `sysctl` osx way was lost, at least in portable environments.
This fixes it, but will not be compatible with OSX > 10.4. 
So i'm sorry but here we loose compatibility with Macs olders than 2005. (🤣)