Added PowerShell fallback for WSL
If WSLENV is not set (which, for my clean installation of Win10 with Debian,
it was not), PATH is not populated with Windows' entries and is missing
PowerShell executable, causing the COM port listing to fail.

This commit adds a fallback with the default path if not found in PATH, and
also replaces the wmic.exe check (which is no longer used) with a PSH check.