armsrc/felica: add polling response time slot support to felica_sim_lite
Make sure we respond in a valid time slot.
We were responding too early, the first time slot (which is always allowed) starts after 512 * 64 / fc.

Also, respond in incremental time slots. We could have used a fixed time slot (#0 is always allowed), but by using incremental time slots we have a better chance the reader will get our response in the expected time frame.
(when our timing is wrong, we have a risk of responding too early when using the first slot, and too late when using the last slot. Slots in the middle of the allowed range have the best chance)