First round of refactoring for Fudan
Went through and started improving the Fudan code.

- the dump command now can take -v or --verbose
- began refactoring the dump command (Still needs more work but it's stable)
- Added a few other Fudan chips to the list (Not detectable yet because of not having any to get the adqa or sak for them, but once we can, they should just work:tm:)

With these changes, the view command would no longer work since the files that dump make are different than what they used to be. I disabled it with a warning for now until the dump command is fully reimplemented and the view command can take the block size and number of sectors into account.

Signed-off-by: JLitewski <hackhalotwo@gmail.com>