Fix SimulateIso14443aInit usage
Added NULL as `uint8_t *iRats` to all invokations in armsrc/Standalone.

Not sure if that is the right thing to do: The actual code checks for the presence of a special flag, but does not contain a check for NULL.
IMHO adding a NULL check in `iso14443a.c` line 1259 wouldn't hurt, and could catch programming errors (flag set by accident, but NULL passed).

Maybe @evildaemond wants to verify whether my fix does the right thing or if I missed something (they added the iRATs to `iso14443a.h` with 56324b16b2).