Updated hf iclass configcard
Documented and added an extensive list of configcards available and functionalities.
- Led Operations (A8 byte) are now fully documented (up from 2)
- Added : Beep Operations
- Elite Key Operations fully documented (3 up from 2)
-Added potential other bytes pending full documentation / discovery
-To Add: Reader Master Key Operations and Erratic Operations

While initially I was trying to maximise backward compatibility with cardhelper, the way its functions were linked and engrained in the code, caused a constant flow of errors and did not allow the flexibility to easily expand the list of cards. Maintaining that code would lead to creating multiple duplicate functions with no real added value perceived (except increased complexity of the code and difficulty to continue maintaining its  backward compatiblity)